Bug correction: the result of the median was sometimes wrong if the addition of some variables was overflowing an INT

+++ algorithm/get_median.c
@@ -6,7 +6,7 @@
 /*   By: sdummett <sdummett@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/30 17:08:00 by sdummett          #+#    #+#             */
-/*   Updated: 2021/09/04 16:30:25 by sdummett         ###   ########.fr       */
+/*   Updated: 2021/09/04 18:29:29 by sdummett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,14 +18,14 @@
 
 int	get_median(t_sorted *sorted)
 {
-	unsigned int			index_median;
-	int						median;
+	unsigned int	index_median;
+	long int		median;
 
 	index_median = (sorted->total + 1) / 2;
 	if (index_median % 2 == 0)
-		median = sorted->list[index_median - 1];
+		median = (long int)sorted->list[index_median - 1];
 	else
-		median = (sorted->list[index_median - 1] + \
-		sorted->list[index_median]) / 2;
-	return (median);
+		median = ((long int)sorted->list[index_median - 1] + \
+		(long int)sorted->list[index_median]) / (long int)2;
+	return ((int)median);
 }
